+++ packages/genai_primitives/lib/src/parts/model.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-/// @docImport 'part.dart';
+/// @docImport 'standard_part.dart';
 library;
 
 import 'dart:convert';
@@ -25,9 +25,9 @@ import 'package:meta/meta.dart';
 /// * Pass extended [defaultPartConverterRegistry] to all methods `fromJson`
 ///   that accept a converter registry.
 @immutable
-abstract base class BasePart {
+abstract base class Part {
   /// Creates a new part.
-  const BasePart();
+  const Part();
 
   /// The key of the part type in the JSON representation.
   static const String typeKey = 'type';
@@ -35,7 +35,7 @@ abstract base class BasePart {
   /// Deserializes a part from a JSON map.
   ///
   /// The [converterRegistry] parameter is a map of part types to converters.
-  factory BasePart.fromJson(
+  factory Part.fromJson(
     Map<String, Object?> json, {
     required Map<String, JsonToPartConverter> converterRegistry,
   }) {
@@ -55,14 +55,14 @@ abstract base class BasePart {
   Map<String, Object?> toJson();
 }
 
-typedef JsonToPartConverter<T extends BasePart> =
+typedef JsonToPartConverter<T extends Part> =
     Converter<Map<String, Object?>, T>;
 
 typedef _JsonToPartFunction<T> = T Function(Map<String, Object?> json);
 
-/// A converter that converts a JSON map to a [BasePart].
+/// A converter that converts a JSON map to a [Part].
 @visibleForTesting
-class PartConverter<T extends BasePart> extends JsonToPartConverter<T> {
+class PartConverter<T extends Part> extends JsonToPartConverter<T> {
   const PartConverter(this._function);
 
   final _JsonToPartFunction<T> _function;